Configure Maven
Changes the Maven build so that it won't run tests in the SlowWrapperTest
category. For an unknown reason, these tests are non-determnistic, and
fail randomly with too many or few interactions. Until they're fixed, they
can't be run as a part of the build. In any case they're slow, and more
like integration than unit tests. It's still worth keeping them around, so
that they can be run manually to verify functionality.