Return persistent copy of SharedTable
getTable() now returns a reference to a copy of the singleton Table. It returns a reference to the same object on every call,
because otherwise using the table as an output @Parameter of a plug-in wouldn't work correctly. For example, if getTable()
created a new object on each wall, then running ConnectivityWrapper and AnisotropyWrapper would open two tables with separate results instead
of one with all the values.