Update to pom-scijava 24.0.0
This corresponds to imagej 2.0.0-rc-69.

And make the necessary changes to account for backwards
incompatibilities in the API. Specifically, the net.imagej.table
package moved to org.scijava.table in a new scijava-table component.
Unfortunately, it did not end up being feasible to preserve the old
package deprecated, so usages must all be updated in one swoop.