remove scale from EF (and findRidgeTest)
EF does not need to know the image scale.
This commit removes the unnecessary variable and computations associated with the scale.
The computation now exists entirely at a (continuous or discrete) integer pixel scale.
This reduces computation time for ellipsoid optimisation from 30 s to 20 s for emu with skipRatio 10.

also, changed around the FindRidgePoints test: for some weird reason "mvn clean test" was taking ages for this simple test.
The contents of the test did not change.