Enhance Anisotropy to give consistent results
This commit introduces a new utility for visualising data,
which users may find helpful to understand their anisotropy results

The default sampling increment is set to 2.3 × pixel spacing,
which is about where the pixel spacing no longer creates an
'imprint' on the MIL of the structure's texture.
If sampling increment is too small (≲ 2px), the pixel grid size starts
to interfere with the feature size.

Count every foreground/background boundary
following the definition in Odgaard 1997, and the use in BoneJ1

Print the eigenvalues to the results table, after a user request

Update test to account for change in boundary counting logic