Fix UsageReporter in tests
In most tests it's enough to mock UsageReporter once for all tests,
which can be done in @BeforeClass. In AnalySkeletonWrapper tests the
behaviour of UsageReporter is actually tested. There we can create
the mock once, and then just reset the behaviour & state after each
tests, so that individual test starts with a clean slate.