Anisotropy style fixes (#205)
* Check constructor parameters

* Fix naming

* Remove "seed" parameters

Removes the option to set the seed of the PRNGs in both the MIL op and
wrapper. The parameters weren't used for anything.

* Mark static test resources for gc

I read yesterday about memory leaks in Java, and one potential source is
big static objects. They can't be garbage collected until a class is
unloaded. Thus I'm marking such references null in the test classes.