Fix line generation bias
Last night when I couldn't sleep I realised there was a slight bias in
the line generation of MIL. For example, let's assume that you generated lines from the
four quadrants of the plane, and called for 7 lines. Because the order
of the quadrants was always the same, there was never a chance that the
bottom-right quadrant might have 2 points.