speed up sampling on ellipsoid surface
150 random vectors and numbers are preallocated and refreshed only every 100th step
these "magic" quantities (150,100) were optimised for emu.tif on a 4-core laptop
this might reflect a typical use case.

Number of required random vectors and number may vary depending on ellipsoid shape and chance.
This is because the sampling algorithm has an accept/reject step that depends on ellipsoid shape and chance.