Remove messages from IJ2 style dialog boxes
Messages are handled really badly when plugins are called from macros,
and usually block execution until the user clicks OK. There is a hacky
workaround, but it requires users to dive into Java source code to find
the variable name of the String which is presented in the message (which
is way way too much to ask of most scripting users), and then to escape
it by presenting an empty string in the macro command.

Until @Parameters with ItemVisibility.MESSAGE are handled better we will
avoid them.

Fixes #267