First go at sequential read optimisation (buggy)
This is an attempt to increase array read speed by reading all label
image pixels in 4 stripes, one stripe for each pixel in the
neighbourhood front. So far the overall run time is shorter by c. 20% on
larger images, but the labelling is broken (all particles are given the
same label). It's possible that the speedup is due not to faster label
array reads but to reduced handling of the label Map, which itself is an
expensive random array access.