First go at a basic brute force ray-caster
This class generates a list of points that are visible from each
skeleton point. It is brute force, checking for each foreground pixel,
whether each skeleton point is visible. An optimisation is included to
not check foreground pixels that are not on the surface (i.e. have no
background neighbours).

It may be used to provide a point cloud for each skeleton point, against
which ellipsoid optimisation can be performed, instead of the current
method of checking collisions between points generated on an ellipsoid
and the image pixels. This centroid+point cloud is a small amount of
data that could be uploaded to GPU for faster optimisation.