Basic event logged with GA4 but many things missing/broken
Web report screen needs quite a bit of configuration and is slow and
clunky compared to old GA. Consider abandoning GA if the new service is
not better than the old service.

Security consideration: this commit includes an API secret and a
measurement ID, which could be used to add fake event data to the
account (but it's not obvious how these tokens could be used to get
access to the web report interface or to do anything other than log
events). Obscuring the account data seems futile because the code could
easily be re-written by anyone to make the decrypted secrets visible in
plaintext. While acknowledging the bad practice of hardcoding account
details, there doesn't seem much point in putting the effort into not
hardcoding them.