Use dedicated 3-way sort (#339)
* Use dedicated 3-way sort

Arrays.sort carries overhead because it's generalised. But we know
precisely the case here, which is always a 3-element array. Profiling
shows that while assigning ellipsoid IDs most (98% !) of the time is
spent in getSortedRadii(). This dedicated 3-way sort is about 2× faster
than Arrays.sort in a basic performance test.

* Reduce slow methods a bit more

* getSortedRadii() and Math.max() are similar speed

But the result is needed twice (for min and max) so getSortedRadii() is
better here.

* formatting