Implement a bounding box
A bounding box that includes visible skeleton points and excludes
invisible ones is implemented. It relies on an initial sample of
skeleton points then finds the min and max extremes in x, y, z that
contain the visible skeleton points and the next nearest occluded
skeleton points. The intention is that for large volumes most of the
skeleton points are outside the bounding box and are skipped for ray
tracing.