Clear and set JOCL device names in Prefs
This sets the Prefs to record the devices visible to JOCL at the time
the GPU Check was run.

This is a basis to detect unexpected changes to GPUs. Later programs
can check whether the name of the device they see is consistent with
the device(s) selected for use.