Prepare for first BoneJ+ release, with bonej-plus 0.0.3 (#358)
* Display an AWT-style dialog that lists all GPUs

Compliant GPUs are selected by default, and non-compliant GPUs are
non-selectable and greyed out.

* Make GPU preferences persist using IJ.Prefs

* Use the words "platform" and "device" in the GPU checker GUI

* Disambiguate device strings for scriptability

* Make sure that non-compliant devices are greyed out and unchecked.

Even when use all devices is selected.

* Improve handling of GPU selection

* Revert "Comment out dependency on bonej-plus for now"

This reverts commit 0815e64efd503367bd0b39c96b8d275ee43b797e.

* use the new purgeAll method to clean up all the BoneJ+ classes.

* Clear and set JOCL device names in Prefs

This sets the Prefs to record the devices visible to JOCL at the time
the GPU Check was run.

This is a basis to detect unexpected changes to GPUs. Later programs
can check whether the name of the device they see is consistent with
the device(s) selected for use.

* Improve the prefs checking logic and also clear useDevice entries

* Simplify GPU checker

GPU checker resets Prefs each time it is run, and sets each compliant
device to be used.

* Initialise the random number generators only when needed

It's common that a QuckEllipsoid could be instantiated without needing
the surface point generating functions that require RNG or their
buffers of random numbers. So leave them null until they are needed.

* Don't run the GPU Checker if a regulated plugin is running

* Add logback configuration files to Modern/*/src/main/resources

* Use first soft release version of BoneJ+ (0.0.1)

* Update copyright blurbs

* Use bonej-plus-0.0.2

* Set logback level to ERROR

* Make GPUCheckerWrapper extend BoneJCommand

This makes it reappear in the menu

* Remove logback entirely

No idea what it was meant to be doing, but it's super annoying.
fixes #349

* Bump bonej-plus to 0.0.3-SNAPSHOT

* GPU Checker needs a BeanCounter instance

* Prepare to release with bonej-plus-0.0.3

* Use default Scijava version of Mockito (4.11.0)

Was holding version at 4.11.0 until Scijava caught up.

* Use the bonej-plus GitHub Maven package repository

* Try simpler PR build to test authentication

* Revert "Try simpler PR build to test authentication"

This reverts commit ac0ab07aee54a55c0f600bf4664b55d662ce1884.

* Try maven-settings-actions to authenticate to get package

* Add dependency on JOCL to bonej-plugins

* Use github-settings-action to authenticate